+++ src/Hevadea/Entities/Components/Light.cs
@@ -1,12 +1,13 @@
 ﻿using Microsoft.Xna.Framework;
 
-namespace Hevadea.Entities.Components.States
+namespace Hevadea.Entities.Components
 {
     public class LightSource : EntityComponent
     {
-        public bool IsOn { get; set; } = false;
-        public bool IsOff => !IsOn;
         public int Power { get; set; } = 32;
         public Color Color { get; set; } = Color.White;
+
+        public bool IsOn { get; set; } = false;
+        public bool IsOff => !IsOn;
     }
 }