+++ src/Hevadea/Entities/Components/Physic.cs
@@ -10,7 +10,7 @@ namespace Hevadea.Entities.Components.Attributes
     public class Physic : EntityComponent
     {
         public float Mass { get; set; } = 1f;
-        public Vector2 Speed { get; set; } = new Vector2();
+        public Vector2 Velocity { get; set; } = new Vector2();
         public Vector2 Acceleration { get; set; } = new Vector2();
     }
 }