+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - The save folder move from the application directory to "Documents/My Games/Hevadea/Saves".
 
 ### Game
+- **Added** lantern Items, Entity, Recipe: light source that can be carry by the player.
 
 ### Framework
 - Sound support